Renaming Philomena and scraping-related classes directory
