+++ src/lib/philomena/search/QueryLexer.ts
